// EPS-Client/src/app/modules/api/model/projectResponse.ts
/**
 * Agency Pulse API v1.0
 * API für das Agency Pulse - Einsatzplanungssystem
 *
 * OpenAPI spec version: 1.0
 * Contact: development@de.cpm-int.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttributeResponse } from './attributeResponse';
import { CategoryResponse } from './categoryResponse';
import { ClientResponse } from './clientResponse';
import { PersonResponse } from './personResponse';
import { QuestionnaireResponse } from './questionnaireResponse';
import { StatusResponse } from './statusResponse';

export interface ProjectResponse { 
    id?: string;
    parentId?: string;
    typeId?: number;
    name?: string;
    number?: string;
    position?: number;
    from?: Date;
    to?: Date;
    usable?: boolean;
    editable?: boolean;
    client?: ClientResponse;
    questionnaires?: Array<QuestionnaireResponse>;
    categories?: Array<CategoryResponse>;
    appointmentAttributes?: Array<AttributeResponse>;
    status?: Array<StatusResponse>;
    persons?: Array<PersonResponse>;
}